+++ src/main/java/quenfo/de/uni_koeln/spinfo/core/application/Launcher.java
@@ -3,11 +3,13 @@ package quenfo.de.uni_koeln.spinfo.core.application;
 import java.io.IOException;
 import java.sql.SQLException;
 
+import org.apache.log4j.Logger;
+
 import quenfo.de.uni_koeln.spinfo.classification.applications.ClassifyDatabase;
-import quenfo.de.uni_koeln.spinfo.information_extraction.applicationsjb.ExtractNewCompetences;
-import quenfo.de.uni_koeln.spinfo.information_extraction.applicationsjb.ExtractNewTools;
-import quenfo.de.uni_koeln.spinfo.information_extraction.applicationsjb.MatchCompetences;
-import quenfo.de.uni_koeln.spinfo.information_extraction.applicationsjb.MatchTools;
+import quenfo.de.uni_koeln.spinfo.information_extraction.applications.ExtractNewCompetences;
+import quenfo.de.uni_koeln.spinfo.information_extraction.applications.ExtractNewTools;
+import quenfo.de.uni_koeln.spinfo.information_extraction.applications.MatchCompetences;
+import quenfo.de.uni_koeln.spinfo.information_extraction.applications.MatchTools;
 
 /**
  * Class contains an application to run different workflows from command line
@@ -16,23 +18,30 @@ import quenfo.de.uni_koeln.spinfo.information_extraction.applicationsjb.MatchToo
  *
  */
 public class Launcher {
+	
+	static Logger log = Logger.getLogger(Launcher.class);
 
 	public static void main(String[] args) throws ClassNotFoundException, SQLException, IOException {
 		
 		switch (args[0]) {
 		case "classify":
+			log.info("--- Classification starts ---");
 			ClassifyDatabase.main(args);
 			break;
 		case "extractCompetences":
+			log.info("--- Extracting competences starts ---");
 			ExtractNewCompetences.main(args);
 			break;
 		case "extractTools":
+			log.info("--- Extracting tools starts ---");
 			ExtractNewTools.main(args);
 			break;
 		case "matchCompetences":
+			log.info("--- Matching competences starts ---");
 			MatchCompetences.main(args);
 			break;
 		case "matchTools":
+			log.info("--- Matching tools starts ---");
 			MatchTools.main(args);
 			break;
 		default:
